+++ substrafl/dependency/__init__.py
@@ -1,0 +1,7 @@
+from substrafl.dependency.constants import EXCLUDED_PATHS_REGEX_DEFAULT
+from substrafl.dependency.schemas import Dependency
+
+__all__ = [
+    "Dependency",
+    "EXCLUDED_PATHS_REGEX_DEFAULT",
+]