+++ CHANGELOG.md
@@ -76,10 +76,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enforce kwargs for user facing function with more than 3 parameters ([#109](https://github.com/Substra/substrafl/pull/109))
 - Remove references to `composite`. Replace by `train_task`. ([#108](https://github.com/Substra/substrafl/pull/108))
+- Way to copy function files ([#118](https://github.com/Substra/substrafl/pull/118))
 
 ### Added
 
 - Add the Federated Principal Component Analysis strategy ([#97](https://github.com/Substra/substrafl/pull/97))
+- Add default exclusion patterns when copying file to avoid creating large Docker images ([#118](https://github.com/Substra/substrafl/pull/118))
 
 ## [0.35.1](https://github.com/Substra/substrafl/releases/tag/0.35.1) - 2023-04-11